Add files via upload
This document currently opens a list of articles from the New York Times that contain "Ada Louise Huxtable." It finds the author of the article (most of the time, occasionally letters to the editor are stored in funky ways in the NYT archives) and checks to see if the author string found contains "Ada Louise Huxtable" (so papers she authors with others are included). Then, the program tries to find the date and year at each article. For some reason, one type of article (the one that prints "=> found here! in the author position) does not return its date and year properly, even though this information is stored the same as other articles that properly return the date and year.